app/src/main/java/com/example/forest/repository/provider/EncryptedSharedPreferencesProvider.kt
package com.example.forest.repository.provider

import android.content.Context
import android.content.SharedPreferences
import androidx.annotation.VisibleForTesting
import androidx.security.crypto.EncryptedSharedPreferences
import androidx.security.crypto.MasterKeys
import javax.inject.Provider

@VisibleForTesting
const val SECURE_STORAGE_FILENAME = "sharedPrefSettings"

class EncryptedSharedPreferencesProvider(
    private val context: Context
) : Provider<SharedPreferences> {

    override fun get(): SharedPreferences {
        val keyGenParameterSpec = MasterKeys.AES256_GCM_SPEC
        val masterKeyAlias = MasterKeys.getOrCreate(keyGenParameterSpec)
        return EncryptedSharedPreferences.create(
            SECURE_STORAGE_FILENAME,
            masterKeyAlias,
            context.applicationContext,
            EncryptedSharedPreferences.PrefKeyEncryptionScheme.AES256_SIV,
            EncryptedSharedPreferences.PrefValueEncryptionScheme.AES256_GCM
        )
    }
}